Do app (dll) harness entrypoints for apps
This permits to run common code on the dll side before calling the
app-land code.
With this now we can do common stuff like init imgui/imguizmo, and later
do more "engine" stuff as well.

This is basically the "engine" part of the app, but running in the
reloadable part of the DLL.

We also move some imguizmo stuff from learn_opengl into common entity code