[entity] Make separate entity_manager unit so it can be one struct
Before we had to make components holder a pointer, which required an
allocation, which complicated the setup if we ever we want to just copy
the entity manager.

Now with this separate translation unit, normal entity code can get the
API but only things that care about the size of the EntityManager can
get the other header.